intel: Fix insufficient integer width for upload buffer offset
I was being overly miserly and gave the offset of the buffer into the bo
insufficient bits, distracted by the adjacency of the buffer[4096].

Ref: https://bugs.freedesktop.org/show_bug.cgi?id=34541

+++ src/mesa/drivers/dri/intel/intel_context.h
@@ -201,8 +201,8 @@ struct intel_context
    struct {
       drm_intel_bo *bo;
       GLuint offset;
-      uint16_t buffer_len;
-      uint16_t buffer_offset;
+      uint32_t buffer_len;
+      uint32_t buffer_offset;
       char buffer[4096];
    } upload;
